requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
pydantic>=2.5,<3
PyJWT>=2.8,<3
cryptography>=42,<43
SQLAlchemy[asyncio]>=2.0,<3
aiosqlite>=0.20,<1
qrcode[pil]>=8,<9
# DEPENDENCIAS OCR
pytesseract>=0.3.10
python-multipart>=0.0.9
# DEPENDENCIAS FRONTEND
jinja2>=3.1.0